Fixes #396.
This PR makes it so that both the `TextProgressMeter` and the `TqdmProgressMeter` both work through the progress bar manager. In particular this means that we will not get any printout on the backward pass for the `TextProgressMeter` (as its `meter_idx` is deleted during `close`, called at the end of the forward pass), whereas previously we would get a binomial printout as the backward pass worked through its recursive checkpointing. This means we now have consistency with `TqdmProgressMeter`, for which there is less of a clear way for any providing any possible out during the backward pass.

(I'd be open to changing the above -- adding options for printout during backpropagation -- but this would probably be fairly tricky. I could see it requiring defining a new primitive, called at the start of integration, for closing the "backward bar", whilst still remaining JVP-compatible.)

In addition, we now have a `try`/`except` around the bar lookup during `step`. This means we do not raise an error when the bar does not exist during backpropagation.

Depends on https://github.com/patrick-kidger/equinox/pull/697.